transactions

+++ main/routes.py
@@ -87,8 +87,6 @@ def seller_page():
         if current_user.units >= units:
             with app.app_context():
                 order = SellOrder(user_id=current_user.id, units=units, price=price)
-                seller: User = User.query.filter_by(id=order.user_id).first()
-                seller.units -= units
                 db.session.add(order)
                 db.session.commit()
         else:
@@ -103,18 +101,8 @@ def seller_page():
 def buyer_page():
     if not current_user.is_authenticated:
         return redirect(url_for("login"))
-    form = SellOrderForm()
-    if form.validate_on_submit():
-        units, price = form.unit.data, form.price.data
-        with app.app_context():
-            order = SellOrder(user_id=current_user.id, units=units, price=price)
-            buyer: User = User.query.filter_by(id=current_user.id).first()
-            buyer.units += units
-            db.session.add(order)
-            db.session.commit()
-
     sell_orders = get_sell_orders()
-    return render_template("buyer_page.html", title="Buy Units", sell_orders=sell_orders, form=form)
+    return render_template("buyer_page.html", title="Buy Units", sell_orders=sell_orders)
 
 
 @app.route("/cancel_sell_order", methods=["GET", "POST"])
@@ -128,11 +116,9 @@ def cancel_sell_order():
 
     with app.app_context():
         order: SellOrder = SellOrder.query.filter_by(id=order_id).first()
-        buyer: User = User.query.filter_by(id=order.user_id).first()
-        buyer.units += order.units
         db.session.delete(order)
         db.session.commit()
-    flash(f"Order: {order.id} Closed")
+        flash(f"Order: {order.id} Closed")
     return redirect(url_for('seller_page'))
 
 
@@ -154,14 +140,24 @@ def checkout_page():
             with app.app_context():
                 order: SellOrder = SellOrder.query.filter_by(id=order_id).first()
                 buyer: User = User.query.filter_by(id=current_user.id).first()
-                buyer.units += form.units.data
-                order.units -= form.units.data
-                if order.units == 0:
-                    db.session.delete(order)
-                db.session.commit()
-            flash(f"Purchase successful for {form.units.data} units at total {total_price}!")
-            return redirect(url_for('home'))
-        flash('Units exceeded!', 'error')
+                seller: User = User.query.filter_by(id=order.user_id).first()
+                units = form.units.data
+                if seller.units >= units:
+                    if buyer.units + form.units.data <= buyer.capacity:
+                        buyer.units += units
+                        order.units -= units
+                        seller.units -= units
+                        if order.units == 0:
+                            db.session.delete(order)
+                        db.session.commit()
+                        flash(f"Purchase successful for {form.units.data} units at total {total_price}!")
+                        return redirect(url_for('home'))
+                    else:
+                        flash(f"Purchase unsuccessful for {form.units.data} units. Your Battery is Full!")
+                else:
+                    flash(f"Purchase unsuccessful for {form.units.data} units. Seller Does Not have these Units!")
+        else:
+            flash('Units exceeded!', 'error')
     return render_template("checkout_page.html", title="Checkout", order=order, seller=seller, form=form)
